Move worker files in to `worker` package
