+++ app/src/main/java/com/keylesspalace/tusky/components/notifications/NotificationHelper.java
@@ -63,6 +63,7 @@ import com.keylesspalace.tusky.entity.Status;
 import com.keylesspalace.tusky.receiver.SendStatusBroadcastReceiver;
 import com.keylesspalace.tusky.util.StringUtils;
 import com.keylesspalace.tusky.viewdata.PollViewDataKt;
+import com.keylesspalace.tusky.worker.NotificationWorker;
 
 import java.util.ArrayList;
 import java.util.Collections;
