+++ app/src/main/java/com/keylesspalace/tusky/receiver/UnifiedPushBroadcastReceiver.kt
@@ -20,11 +20,11 @@ import android.content.Intent
 import android.util.Log
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkManager
-import com.keylesspalace.tusky.components.notifications.NotificationWorker
 import com.keylesspalace.tusky.components.notifications.registerUnifiedPushEndpoint
 import com.keylesspalace.tusky.components.notifications.unregisterUnifiedPushEndpoint
 import com.keylesspalace.tusky.db.AccountManager
 import com.keylesspalace.tusky.network.MastodonApi
+import com.keylesspalace.tusky.worker.NotificationWorker
 import dagger.android.AndroidInjection
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
