+++ app/src/main/java/com/keylesspalace/tusky/worker/NotificationWorker.kt
@@ -1,24 +1,26 @@
-/* Copyright 2020 Tusky Contributors
+/*
+ * Copyright 2023 Tusky Contributors
  *
- * This file is part of Tusky.
+ * This file is a part of Tusky.
  *
- * Tusky is free software: you can redistribute it and/or modify it under the terms of the GNU
- * Lesser General Public License as published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version.
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 3 of the
+ * License, or (at your option) any later version.
  *
  * Tusky is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
- * General Public License for more details.
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
  *
- * You should have received a copy of the GNU Lesser General Public License along with Tusky. If
- * not, see <http://www.gnu.org/licenses/>. */
+ * You should have received a copy of the GNU General Public License along with Tusky; if not,
+ * see <http://www.gnu.org/licenses>.
+ */
 
-package com.keylesspalace.tusky.components.notifications
+package com.keylesspalace.tusky.worker
 
 import android.content.Context
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.keylesspalace.tusky.ChildWorkerFactory
+import com.keylesspalace.tusky.components.notifications.NotificationFetcher
 import javax.inject.Inject
 
 /** Fetch and show new notifications. */
